+++ src/index.ts
@@ -100,8 +100,9 @@ interface AlgorithmArgs {
 
 
 class TheAlgorithm {
+    static isDebugMode = isDebugMode;
+
     filters: FeedFilterSettings = buildNewFilterSettings();
-    isDebug = isDebugMode;
     lastLoadTimeInSeconds: number | null = null;  // Duration of the last load in seconds
     // TODO: loadingStatus has become the main flag for whether the feed is loading or not. Not great.
     loadingStatus: string | null = READY_TO_LOAD_MSG;  // String describing load activity (undefined means load complete)