+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Make TheAlgorithm.isDebugMode a static
 
 ### v0.39.8
 * More futzing with type exports